Allow better control of browser sessions when using multiple actors
You can now use the `serenity.different.browsers.for.each.actor` (set to true by default) to make multiple actors use the same browser, even in the same test. This can be useful if actors are used to illustrate the intent of a test, but no tests use more than one actor simultaneously.