chore: allow customization of default guice bindings
Implement `Injectors.setDefaultModule(Module)` to allow setting custom guice binding for the default injector (retrieved via `Injectors.getInjector()`).

Why is this required: The current guice bindings define many singletons. When the serenity-gradle-plugin makes use of these bindings, the singletons are loaded into memory, and because gradle is started by default as a daemon and keeps running in the background, the singletons don't get updated on subsequent invocation of gradle tasks.